Ensure backward compatibility wrt old config
