+++ src/main/java/org/cloudfoundry/promregator/config/AuthenticatorConfiguration.java
@@ -5,7 +5,10 @@ public class AuthenticatorConfiguration {
 	
 	private BasicAuthenticationConfiguration basic;
 
-	private OAuth2XSUAAAuthenticationConfiguration oauth2xsuaa;
+	//@Deprecated
+	private OAuth2XSUAABasicAuthenticationConfiguration oauth2xsuaa;
+
+	private OAuth2XSUAABasicAuthenticationConfiguration oauth2xsuaaBasic;
 
 	private OAuth2XSUAACertificateAuthenticationConfiguration oauth2xsuaaCertificate;
 
@@ -25,12 +28,22 @@ public class AuthenticatorConfiguration {
 		this.basic = basic;
 	}
 
-	public OAuth2XSUAAAuthenticationConfiguration getOauth2xsuaa() {
+	//@Deprecated
+	public OAuth2XSUAABasicAuthenticationConfiguration getOauth2xsuaa() {
 		return oauth2xsuaa;
 	}
 
+	//@Deprecated
 	public void setOauth2xsuaa(OAuth2XSUAAAuthenticationConfiguration oauth2xsuaa) {
 		this.oauth2xsuaa = oauth2xsuaa;
+	}
+
+	public OAuth2XSUAABasicAuthenticationConfiguration getOauth2xsuaaBasic() {
+		return oauth2xsuaaBasic;
+	}
+
+	public void setOauth2xsuaaBasic(OAuth2XSUAABasicAuthenticationConfiguration oauth2xsuaaBasic) {
+		this.oauth2xsuaaBasic = oauth2xsuaaBasic;
 	}
 
 	public OAuth2XSUAACertificateAuthenticationConfiguration getOauth2xsuaaCertificate() {
