+++ src/test/java/org/cloudfoundry/promregator/config/OAuth2XSUAAAuthenticationConfigurationTest.java
@@ -10,42 +10,42 @@ public class OAuth2XSUAAAuthenticationConfigurationTest {
 
 	@Test
 	public void testTokenServiceURLBackwardCompatibility() {
-		OAuth2XSUAAAuthenticationConfiguration subject = new OAuth2XSUAAAuthenticationConfiguration();
+		OAuth2XSUAABasicAuthenticationConfiguration subject = new OAuth2XSUAABasicAuthenticationConfiguration();
 		subject.setTokenServiceURL("https://example.org/oauth/token");
 		assertThat(subject.getTokenServiceURL(), equalTo("https://example.org/oauth/token"));
 	}
 
 	@Test
 	public void testTokenServiceURLBackwardCompatibilityWithPrefixInPath() {
-		OAuth2XSUAAAuthenticationConfiguration subject = new OAuth2XSUAAAuthenticationConfiguration();
+		OAuth2XSUAABasicAuthenticationConfiguration subject = new OAuth2XSUAABasicAuthenticationConfiguration();
 		subject.setTokenServiceURL("https://example.org/v1/oauth/token");
 		assertThat(subject.getTokenServiceURL(), equalTo("https://example.org/v1/oauth/token"));
 	}
 
 	@Test
 	public void testTokenServiceURLBackwardCompatibilityWithPort() {
-		OAuth2XSUAAAuthenticationConfiguration subject = new OAuth2XSUAAAuthenticationConfiguration();
+		OAuth2XSUAABasicAuthenticationConfiguration subject = new OAuth2XSUAABasicAuthenticationConfiguration();
 		subject.setTokenServiceURL("https://example.org:1234/oauth/token");
 		assertThat(subject.getTokenServiceURL(), equalTo("https://example.org:1234/oauth/token"));
 	}
 
 	@Test
 	public void testScopesWithBlanksAsSeparator() {
-		OAuth2XSUAAAuthenticationConfiguration subject = new OAuth2XSUAAAuthenticationConfiguration();
+		OAuth2XSUAABasicAuthenticationConfiguration subject = new OAuth2XSUAABasicAuthenticationConfiguration();
 		subject.setScopes("a b c");
 		assertThat(subject.getScopes(), containsInAnyOrder("a", "b", "c"));
 	}
 
 	@Test
 	public void testScopesWithCommaAsSepartor() {
-		OAuth2XSUAAAuthenticationConfiguration subject = new OAuth2XSUAAAuthenticationConfiguration();
+		OAuth2XSUAABasicAuthenticationConfiguration subject = new OAuth2XSUAABasicAuthenticationConfiguration();
 		subject.setScopes("a, b  , c,d, e");
 		assertThat(subject.getScopes(), containsInAnyOrder("a", "b", "c", "d", "e"));
 	}
 
 	@Test
 	public void testScopesWithMixedSepartors() {
-		OAuth2XSUAAAuthenticationConfiguration subject = new OAuth2XSUAAAuthenticationConfiguration();
+		OAuth2XSUAABasicAuthenticationConfiguration subject = new OAuth2XSUAABasicAuthenticationConfiguration();
 		subject.setScopes("a b, c");
 		assertThat(subject.getScopes(), containsInAnyOrder("a", "b", "c"));
 	}
